e2e/requirements.txt
pytest-playwright==0.7.0
python-dotenv==1.1.0
pytest-dependency==0.6.0
